move `query_callback` to `message_queue.py`
