+++ app/HutchAgent/hutchagent/query.py
@@ -305,68 +305,3 @@ class RQuestQuery:
             .subquery()
         )
         return select(func.count()).select_from(stmt)
-
-
-def query_callback(
-    channel: Channel, method: Basic.Deliver, properties: BasicProperties, body: bytes
-):
-    """The callback to be used when consuming messages from the queue.
-    The arguments to this function will be passed by the channel when a
-    message is consumed.
-
-    Args:
-        channel (Channel): The channel object.
-        method (Deliver): The delivery object.
-        properties (BasicProperties): The message properties.
-        body (bytes): The body of the message.
-    """
-    logger = logging.getLogger(os.getenv("DB_LOGGER_NAME"))
-    response_data = {
-        "protocol_version": "v2",
-        "query_result": dict(),
-    }
-    logger.info("Received message from the Queue. Processing...")
-    try:
-        body_json = json.loads(body)
-        query = RQuestQuery(**body_json)
-        response_data["collection_id"] = query.collection
-        logger.info(f"Successfully unpacked message.")
-    except json.decoder.JSONDecodeError:
-        logger.error("Failed to decode the message from the queue.")
-
-    db_manager = SyncDBManager(
-        username=os.getenv("DATASOURCE_DB_USERNAME"),
-        password=os.getenv("DATASOURCE_DB_PASSWORD"),
-        host=os.getenv("DATASOURCE_DB_HOST"),
-        port=int(os.getenv("DATASOURCE_DB_PORT")),
-        database=os.getenv("DATASOURCE_DB_DATABASE"),
-        drivername=os.getenv("DATASOURCE_DB_DRIVERNAME"),
-    )
-    try:
-        query_start = time.time()
-        res = db_manager.execute_and_fetch(query.to_sql())
-        query_end = time.time()
-        count_ = res[0][0]
-        response_data["query_result"].update(count=count_, status="ok")
-        logger.info(
-            f"Collected {count_} results from query {query.uuid} in {(query_end - query_start):.3f}s."
-        )
-    except sql_exc.NoSuchTableError as table_error:
-        logger.error(str(table_error))
-        response_data["query_result"].update(count=0, status="error")
-    except sql_exc.NoSuchColumnError as column_error:
-        logger.error(str(column_error))
-        response_data["query_result"].update(count=0, status="error")
-    except sql_exc.ProgrammingError as programming_error:
-        logger.error(str(programming_error))
-        response_data["query_result"].update(count=0, status="error")
-
-    try:
-        requests.post(os.getenv("MANAGER_URL"), response_data)
-        logger.info("Sent results to manager.")
-    except req_exc.ConnectionError as connection_error:
-        logger.error(str(connection_error))
-    except req_exc.Timeout as timeout_error:
-        logger.error(str(timeout_error))
-    except req_exc.MissingSchema as missing_schema_error:
-        logger.error(str(missing_schema_error))
